SLING-10031 Move MediaRangeList from Servlets POST to Sling API
This reverts commit dc05de78c8cf6eafa72b2d929b26f9bf7c260789, getting the changes back to the
current release.
