+++ pom.xml
@@ -142,7 +142,7 @@
         <dependency>
             <groupId>org.apache.sling</groupId>
             <artifactId>org.apache.sling.api</artifactId>
-            <version>2.9.0</version>
+            <version>2.23.2</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
